new ubication - improved routines

+++ uninstall.sh
@@ -9,13 +9,22 @@ pid=`pgrep keyboard_leds`
 kill -SIGKILL  $pid
 
 
-# removing all routines and services
-sudo rm /usr/local/bin/keyboard_leds.sh
-sudo rm /usr/local/bin/keyboard_switch.sh
-sudo rm /etc/systemd/system/keyboard_leds.service
+# removing all routines 
+sudo rm -r /opt/led-wayland/
+
+
+# removing Systemd servcie
+if [ -f /etc/systemd/system/keyboard_leds.service ]
+then
+    sudo rm /etc/systemd/system/keyboard_leds.service
+fi
+
 
 # removing user permissions
-sudo rm /etc/sudoers.d/leds_wayland 
+if [ -f /etc/sudoers.d/leds_wayland ]
+then
+    sudo rm /etc/sudoers.d/leds_wayland 
+fi
 
 # reloading services - forcing initialization
 sudo systemctl daemon-reload
